R load user in store (#85)
This feature is Loading the user.status (isLogged or need2fa or error) in RootLayout as a loader.
